split_banlance/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
black==23.10.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
djangorestframework==3.14.0
flake8==6.1.0
isort==5.12.0
kombu==5.3.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
